[refactor]: co-locate Wan VAE and config
