+++ fastvideo/AGENTS.md
@@ -47,6 +47,7 @@ fastvideo/
 | Add a new pipeline class | `pipelines/basic/<model>/` + `configs/pipelines/<model>.py` + register in `registry.py` |
 | Add a new model component | `models/<role>/<model>.py` + `configs/models/<role>/<model>.py` |
 | Edit Wan's dense transformer or arch config | `models/wan/transformer.py` + `models/wan/config.py`; old paths are compatibility shims |
+| Edit Wan's VAE or its config | `models/wan/vae.py` + `models/wan/vae_config.py`; old paths are compatibility shims |
 | Wire an existing model into a new pipeline | `pipelines/basic/<model>/presets.py` + reuse stages from `pipelines/stages/` |
 | Add a converter | `scripts/checkpoint_conversion/<model>_to_*.py` (separate dir, separate AGENTS.md) |
 | Add an attention backend | `attention/backends/<name>.py` + register in selector |
@@ -67,5 +68,6 @@ These dirs are listed in `.pre-commit-config.yaml` `exclude`:
 
 Editing files there will NOT trigger yapf/ruff/mypy/codespell. Format manually if a sibling file shows clear style; do not introduce new violations.
 
-Exception: `models/wan/config.py` and `models/wan/__init__.py` are checked, so
-the family-local config retains its prior lint coverage.
+Exception: `models/wan/config.py`, `models/wan/vae_config.py`, and
+`models/wan/__init__.py` are checked, so the family-local configs retain their
+prior lint coverage.
