+++ fastvideo/configs/AGENTS.md
@@ -4,10 +4,11 @@
 
 Two layers of dataclass configs feed every pipeline: **arch configs** (what the model is) and **pipeline configs** (how to run it).
 
-Wan's architecture config is co-located with its transformer in
-`fastvideo/models/wan/config.py`. `models/dits/wanvideo.py` under this directory
-is a compatibility re-export; pipeline configs remain in `pipelines/wan.py`.
-Other model families retain the layout below.
+Wan's transformer config is co-located in `fastvideo/models/wan/config.py`, and
+its VAE config in `fastvideo/models/wan/vae_config.py`. `models/dits/wanvideo.py`
+and `models/vaes/wanvae.py` under this directory are compatibility re-exports;
+pipeline configs remain in `pipelines/wan.py`. Other model families retain the
+layout below.
 
 ## Layout
