+++ fastvideo/configs/models/vaes/__init__.py
@@ -13,7 +13,7 @@ from fastvideo.configs.models.vaes.minimax_h3_video import (
 )
 from fastvideo.configs.models.vaes.oobleck import OobleckVAEArchConfig, OobleckVAEConfig
 from fastvideo.configs.models.vaes.flux2vae import Flux2VAEConfig
-from fastvideo.configs.models.vaes.wanvae import WanVAEConfig
+from fastvideo.models.wan.vae_config import WanVAEConfig
 
 __all__ = [
     "GameCraftVAEConfig",
