+++ fastvideo/models/wan/__init__.py
@@ -1,2 +1,2 @@
 # SPDX-License-Identifier: Apache-2.0
-"""Wan components; import config or transformer explicitly to avoid import cycles."""
+"""Wan components; import component and config modules explicitly to avoid import cycles."""
